requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==1.11.20
django-celery==3.2.2
django-debug-toolbar==1.11
django-extensions==2.1.6
djangorestframework==3.9.1
kombu==3.0.37
python-dotenv==0.10.1
pytz==2018.9
redis==3.2.0
six==1.12.0
sqlparse==0.2.4
vine==1.2.0
